Fix: result should be double, not int

+++ src/Chapter4/Exercise_solutions/b_multiple_test_p_instant.cpp
@@ -55,7 +55,7 @@ TEST_P(ParameterizedTest, TestMultiply)
     int b = GetParam().second;
 
     // Call your normal function
-    int result = Multiply(a, b);
+    double result = Multiply(a, b);
 
     // Perform assertion
     ASSERT_EQ(a * b, result);
